✨ Load data from Airtable API right in the server
Instead of writing it to a file one time by a separate build script
