+++ src/routes/blog/_posts.js
@@ -1,12 +1,3 @@
-const filePath = 'data/posts.json';
-let fs = require('fs');
-let md = require('markdown-it')('commonmark');
+const blogTable = require('../../consumers/Airtable');
 
-export default async () => {
-	return await new Promise((resolve, reject) => {
-		fs.readFile(filePath, 'utf8', (err, data) => {
-			if (err) reject(err);
-			resolve(data);
-		})
-	});
-};
+export default blogTable;
